Refactor compiler phases
Amp-Thread-ID: https://ampcode.com/threads/T-01a01894-d9e7-7258-953b-e1e6db8f7683
